ml_service/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# ML and Computer Vision
tensorflow==2.15.0
opencv-python==4.8.1
numpy==1.24.3
pillow==10.1.0

# Data processing
redis==5.0.1
pydantic==2.4.2

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0

# Monitoring and logging
python-json-logger==2.0.7
prometheus-client==0.19.0
